refactor: Move `ItemSyncError` class to `errors` directory
