+++ src/content/errors/index.ts
@@ -1,2 +1,3 @@
+export { ItemSyncError } from './ItemSyncError';
 export { LocalizableError } from './LocalizableError';
 export { MissingPrefError } from './MissingPrefError';
